[Feat]배송조회api구현 (#16)
* feat : 조회 ResponseDto 생성
-Route의 정보 포장을 위한 RouteDto 추가 생성

* feat : 조회 비지니스 로직 생성 도메인 규칙 추가

* feat : Repository fetch join 추가

* fix : NPE문제 해결
- 상항 연산자 사용하여 NULL일 때 NULL값을 넣음

* feat : 조회 api 생성 및 controller 정리

* fix : Hub조회 반환타입 변경...
여기 PR에 넣으면 안되는데... 죄송합니다.

* fix : Hub조회 반환타입 변경...
여기 PR에 넣으면 안되는데... 죄송합니다.

+++ src/main/java/org/gupang/deliveryservice/domain/entity/DeliveryRouteRecords.java
@@ -37,16 +37,16 @@ public class DeliveryRouteRecords extends BaseEntity {
     private String endHubName;
 
     @Column
-    private Double estimatedDistance; //예상거리
+    private int estimatedDistance; //예상거리
 
     @Column
-    private Integer estimatedDuration; //예상 소요 시간
+    private int estimatedDuration; //예상 소요 시간
 
     @Column
-    private Double actualDistance; //실제 거리
+    private int actualDistance; //실제 거리
 
     @Column
-    private Integer actualDuration; //실제 소요 시간
+    private int actualDuration; //실제 소요 시간
 
     @Column
     @Enumerated(EnumType.STRING)
@@ -63,8 +63,8 @@ public class DeliveryRouteRecords extends BaseEntity {
             String startHubName,
             UUID endHubId,
             String endHubName,
-            Double distance,
-            Integer duration,
+            int distance,
+            int duration,
             int sequence) {
             DeliveryRouteRecords routeRecords = new DeliveryRouteRecords();
             routeRecords.routeRecordId = UUID.randomUUID();
